Remove deprecated PowerShell scripts and enhance role assignment cleanup
- Deleted `convert-ambaPack_logs.ps1`, `create-ambalerts.ps1`, `create-pack.ps1`, `get-allmetrics.ps1`, `get-lbmetrics.ps1`, `get-metrics.ps1`, `get-supportedalerts.ps1`, `list-amba-json.ps1`, and `list-amba.ps1` as they are no longer needed.
- Updated `remove-empty-role-assignments.ps1` to include user confirmation before removing role assignments of type 'Unknown'.

+++ tools/remove-empty-role-assignments.ps1
@@ -1,10 +1,25 @@
 # Description: This script will remove any Azure RBAC Role Assignments that have an 'Unknown' Type.
 $OBJTYPE = "Unknown"
-
-#Find and Export-to-CSV Azure RBAC Role Assignments of 'Unknown' Type
-Get-AzRoleAssignment | Where-Object {$_.ObjectType.Equals($OBJTYPE)}| foreach {
-    $object = $_.ObjectId
-    $roledef = $_.RoleDefinitionName
-    $rolescope = $_.Scope
-    Remove-AzRoleAssignment -ObjectId $object -RoleDefinitionName $roledef -Scope $rolescope   
+$ura= Get-AzRoleAssignment | Where-Object {$_.ObjectType.Equals($OBJTYPE)}
+Write-host "Found $($ura.Count) Role Assignments of type $OBJTYPE"
+if ($ura.Count -eq 0) {
+    Write-Host "No Role Assignments of type $OBJTYPE found."
+    exit
+}
+else {
+    #ask for confirmation to remove the role assignments
+    $confirmation = Read-Host "Are you sure you want to remove these role assignments? (Y/N)"
+    if ($confirmation -ne "Y") {
+        Write-Host "Exiting without removing role assignments."
+        exit
+    }
+    else {
+        #Find and Export-to-CSV Azure RBAC Role Assignments of 'Unknown' Type
+        $ura | foreach {
+            $object = $_.ObjectId
+            $roledef = $_.RoleDefinitionName
+            $rolescope = $_.Scope
+            Remove-AzRoleAssignment -ObjectId $object -RoleDefinitionName $roledef -Scope $rolescope   
+        }
+    }
 }
